test(pool): move EC2 pool size tests
